+++ contractors/src/main/java/com/university/contractors/service/AuthenticationService.java
@@ -1,6 +1,5 @@
 package com.university.contractors.service;
 
-import com.university.contractors.config.TokenBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
